+++ src/api/objects/toot.ts
@@ -34,6 +34,7 @@ import {
     MastodonTag,
     MediaCategory,
     StatusList,
+    StringNumberDict,
     TootLike,
     TootScore,
     TrendingLink,
@@ -52,6 +53,12 @@ enum TootVisibility {
 const MAX_ID_IDX = 2;
 const MAX_CONTENT_PREVIEW_CHARS = 110;
 const UNKNOWN = "unknown";
+
+const PROPS_THAT_CHANGE: (keyof Toot)[] = [
+    "favouritesCount",
+    "repliesCount",
+    "reblogsCount"
+];
 
 
 // Extension of mastodon.v1.Status data object with additional properties used by fedialgo
@@ -588,7 +595,21 @@ export default class Toot implements TootObj {
             // Collate multiple retooters if they exist
             let reblogsBy = uriToots.flatMap(toot => toot.reblog?.reblogsBy ?? []);
 
+            const propsThatChange = PROPS_THAT_CHANGE.reduce((props, propName) => {
+                props[propName as string] = Math.max(...uriToots.map(t => (t[propName] as number) || 0));
+                return props;
+            }, {} as StringNumberDict);
+
+            if (uriToots.length > 1) {
+                console.debug(`${logPrefix} propsThatChange: `, propsThatChange);
+            }
+
             uriToots.forEach((toot) => {
+                // Counts may increase over time w/repeated fetches
+                toot.favouritesCount = propsThatChange.favouritesCount;
+                toot.reblogsCount = propsThatChange.reblogsCount;
+                toot.repliesCount = propsThatChange.repliesCount;
+                // booleans can be ORed
                 toot.isFollowed = uriToots.some(toot => toot.isFollowed);
                 toot.muted = uriToots.some(toot => toot.muted);
                 toot.sources = uniquify(uriToots.map(toot => toot.sources || []).flat());